backend/requirements.txt
# backend/requirements.txt

# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / ML — Layer 2 (DistilBERT + FinBERT mock)
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1

# AI / ML — Layer 3 (FAISS dedup mock)
faiss-cpu==1.9.0

# AI / ML — Layer 4 (XGBoost SLA breach predictor mock)
xgboost==2.0.3
scikit-learn==1.4.1.post1

# AI / ML — Layer 5-6 (RAG: ChromaDB + LangChain + Ollama)
langchain==0.1.12
langchain-community==0.0.28
chromadb==0.4.24

# Async tasks
celery==5.3.6
redis==5.0.2

# HTTP
httpx==0.27.0
